fix: исправление замечаний ревью
