+++ src/pages/ingredients/ingredients.module.css
@@ -6,3 +6,7 @@
   align-items: center;
   justify-content: center;
 }
+
+.title {
+  composes:  text text_type_main-large from global;
+}
