+++ src/pages/reset-password/reset-password.jsx
@@ -5,7 +5,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import useInputs from '../../hooks/use-inputs';
 import { Input, PasswordInput, Button } from "@ya.praktikum/react-developer-burger-ui-components";
 import style from "./reset-password.module.css";
-import { resetPassword } from '../../services/actions/actions';
+import { resetPassword } from '../../services/actions/auth-actions';
 
 function ResetPassword() {
   const dispatch = useDispatch();
@@ -36,7 +36,7 @@ function ResetPassword() {
       <h2 className={style.title}>
         Восстановление пароля
       </h2>
-      <form className={style.form}>
+      <form className={style.form} onSubmit={(e) => handleResetPassword(e)}>
         <div className={style.wrapper}>
           <PasswordInput
             placeholder={"Введите новый пароль"}
@@ -59,10 +59,9 @@ function ResetPassword() {
           />
         </div>
         <Button
-          htmlType="button"
+          htmlType="submit"
           type="primary"
           size="medium"
-          onClick={(e) => handleResetPassword(e)}
           disabled={!values.password || !values.token}
         >
           Сохранить
